Move color_extractor services to separate module
