+++ tests/components/color_extractor/test_services.py
@@ -9,7 +9,7 @@ import aiohttp
 import pytest
 from voluptuous.error import MultipleInvalid
 
-from homeassistant.components.color_extractor import (
+from homeassistant.components.color_extractor.services import (
     ATTR_PATH,
     ATTR_URL,
     DOMAIN,
@@ -270,7 +270,9 @@ async def test_file(hass: HomeAssistant, setup_integration) -> None:
     assert state.state == STATE_OFF
 
     # Mock the file handler read with our 1x1 base64 encoded fixture image
-    with patch("homeassistant.components.color_extractor._get_file", _get_file_mock):
+    with patch(
+        "homeassistant.components.color_extractor.services._get_file", _get_file_mock
+    ):
         await hass.services.async_call(DOMAIN, SERVICE_TURN_ON, service_data)
         await hass.async_block_till_done()
 
@@ -305,7 +307,9 @@ async def test_file_denied_dir(hass: HomeAssistant, setup_integration) -> None:
     assert state.state == STATE_OFF
 
     # Mock the file handler read with our 1x1 base64 encoded fixture image
-    with patch("homeassistant.components.color_extractor._get_file", _get_file_mock):
+    with patch(
+        "homeassistant.components.color_extractor.services._get_file", _get_file_mock
+    ):
         await hass.services.async_call(DOMAIN, SERVICE_TURN_ON, service_data)
         await hass.async_block_till_done()
